# complyflow-ai/services/processor/openapi/processor.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplyFlow Processor",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/analyze": {
      "post": {
        "summary": "Analyze",
        "description": "Internal Laravel-to-processor endpoint. Sign the exact request body bytes: HMAC-SHA256(secret, timestamp + LF + nonce + LF + lowercase_hex(SHA256(body))). Send the lowercase hex signature; no trailing LF in the signed message. Timestamp is Unix seconds with an inclusive 60-second clock window. Use a fresh nonce for every attempt, retained for 120 seconds. No browser access. Analysis execution is unavailable until the pipeline is installed.",
        "operationId": "analyze",
        "parameters": [
          {
            "name": "X-CF-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,12}$",
              "title": "X-Cf-Timestamp"
            }
          },
          {
            "name": "X-CF-Nonce",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$",
              "title": "X-Cf-Nonce"
            }
          },
          {
            "name": "X-CF-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$",
              "title": "X-Cf-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid authentication or replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid JSON or closed request schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Authentication or analysis unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Idempotency Key"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementDraft"
            },
            "type": "array",
            "minItems": 1,
            "title": "Requirements"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentDraft"
            },
            "type": "array",
            "minItems": 1,
            "title": "Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "idempotency_key",
          "requirements",
          "documents"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingDraft"
            },
            "type": "array",
            "title": "Findings"
          },
          "processed_documents": {
            "items": {
              "$ref": "#/components/schemas/ProcessedDocumentDraft"
            },
            "type": "array",
            "title": "Processed Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "findings",
          "processed_documents"
        ],
        "title": "AnalyzeResponse"
      },
      "ChunkDraft": {
        "properties": {
          "page_number": {
            "type": "integer",
            "minimum": 1,
            "title": "Page Number"
          },
          "index": {
            "type": "integer",
            "minimum": 0,
            "title": "Index"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Text"
          },
          "start_offset": {
            "type": "integer",
            "minimum": 0,
            "title": "Start Offset"
          },
          "end_offset": {
            "type": "integer",
            "minimum": 0,
            "title": "End Offset"
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 384,
            "minItems": 384,
            "title": "Embedding"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page_number",
          "index",
          "text",
          "start_offset",
          "end_offset",
          "embedding"
        ],
        "title": "ChunkDraft",
        "description": "Page-relative character offsets; end_offset must not precede start_offset."
      },
      "CitationDraft": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "page_number": {
            "type": "integer",
            "minimum": 1,
            "title": "Page Number"
          },
          "quote": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Quote"
          },
          "start_offset": {
            "type": "integer",
            "minimum": 0,
            "title": "Start Offset"
          },
          "end_offset": {
            "type": "integer",
            "minimum": 0,
            "title": "End Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "page_number",
          "quote",
          "start_offset",
          "end_offset"
        ],
        "title": "CitationDraft",
        "description": "Page-relative character offsets; end_offset must exceed start_offset."
      },
      "DocumentDraft": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "content_base64": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Content Base64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "sha256",
          "content_base64"
        ],
        "title": "DocumentDraft"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "FindingDraft": {
        "properties": {
          "requirement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Requirement Id"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "justification": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Justification"
          },
          "confidence": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Confidence"
          },
          "requires_human_review": {
            "type": "boolean",
            "const": true,
            "title": "Requires Human Review"
          },
          "search_summary": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "pattern": "\\S"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Summary"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationDraft"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requirement_id",
          "status",
          "justification",
          "confidence",
          "requires_human_review",
          "search_summary",
          "citations"
        ],
        "title": "FindingDraft",
        "description": "Missing requires no citations and a search summary; other statuses require citations."
      },
      "FindingStatus": {
        "type": "string",
        "enum": [
          "met",
          "partial",
          "missing",
          "inconclusive"
        ],
        "title": "FindingStatus"
      },
      "PageDraft": {
        "properties": {
          "page_number": {
            "type": "integer",
            "minimum": 1,
            "title": "Page Number"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "ocr_used": {
            "type": "boolean",
            "title": "Ocr Used"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page_number",
          "text",
          "ocr_used"
        ],
        "title": "PageDraft"
      },
      "ProcessedDocumentDraft": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageDraft"
            },
            "type": "array",
            "title": "Pages"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkDraft"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "pages",
          "chunks"
        ],
        "title": "ProcessedDocumentDraft"
      },
      "RequirementDraft": {
        "properties": {
          "requirement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Requirement Id"
          },
          "criterion": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Criterion"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Category"
          },
          "weight": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Weight"
          },
          "evaluation_text": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Evaluation Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requirement_id",
          "criterion",
          "category",
          "weight",
          "evaluation_text"
        ],
        "title": "RequirementDraft"
      }
    }
  }
}